+++ callbacks_update_data.py
@@ -1,97 +1,31 @@
 import os
-from dash import Input, Output, callback, ctx, MATCH, State, html, ALL, callback_context
-import dash_uploader as du
+from dash import Input, Output, callback, ctx, State, html, ALL, callback_context
 from app import app, variables
-import utils
-import numpy as np
 import resurfemg.converter_functions as cv
 from pathlib import Path
 from dash.exceptions import PreventUpdate
 from definitions import (PATH_BTN, FILE_PATH_INPUT, STORED_CWD, CWD,
                          CWD_FILES, CONFIRM_CENTERED, MODAL_CENTERED,
                          EMG_OPEN_CENTERED, VENT_OPEN_CENTERED, PARENT_DIR,
-                         LISTED_FILES)
-
-du.configure_upload(app, r"C:\tmp\Uploads", use_upload_id=True)
+                         LISTED_FILES, VENT_FREQUENCY_DIV, VENT_SAMPLING_FREQUENCY, EMG_FREQUENCY_DIV,
+                         EMG_SAMPLING_FREQUENCY)
 
 # variable to keep track of which upload button has been clicked
 clicked_input_btn = None
 
 
-@callback(Output('emg-frequency-div', 'data'),
-          Input('emg-sample-freq', 'value'))
+@callback(Output(EMG_FREQUENCY_DIV, 'data'),
+          Input(EMG_SAMPLING_FREQUENCY, 'value'))
 def update_emg_frequency(freq, ):
     variables.set_emg_freq(freq)
     return 'set'
 
 
-@callback(Output('ventilator-frequency-div', 'data'),
-          Input('ventilator-sample-freq', 'value'))
+@callback(Output(VENT_FREQUENCY_DIV, 'data'),
+          Input(VENT_SAMPLING_FREQUENCY, 'value'))
 def update_ventilator_frequency(freq):
     variables.set_ventilator_freq(freq)
     return 'set'
-
-
-@callback(Output('emg-graphs-container', 'children'),
-          Output('emg-header', 'hidden'),
-          Output('emg-filename', 'children'),
-          Input('hidden-div', 'data'),
-          Input('emg-delete-button', 'n_clicks'))
-def show_raw_data(emg_data, delete):
-    emg_data = variables.get_emg()
-    hidden = True
-
-    trigger_id = ctx.triggered_id
-    filename = variables.get_emg_filename()
-
-    if trigger_id == 'emg-delete-button':
-        variables.set_emg(None)
-        children_emg = []
-    else:
-        if emg_data is not None:
-            emg_frequency = variables.get_emg_freq()
-            children_emg = utils.add_emg_graphs(np.array(emg_data), emg_frequency)
-            hidden = False
-        else:
-            children_emg = []
-
-    return children_emg, hidden, filename
-
-
-@callback(Output('ventilator-graphs-container', 'children'),
-          Output('ventilator-header', 'hidden'),
-          Output('ventilator-filename', 'children'),
-          Input('hidden-div', 'data'),
-          Input('ventilator-delete-button', 'n_clicks'))
-def show_raw_data(ventilator_data, delete):
-    ventilator_data = variables.get_ventilator()
-    hidden = True
-
-    trigger_id = ctx.triggered_id
-    filename = variables.get_ventilator_filename()
-
-    if trigger_id == 'ventilator-delete-button':
-        variables.set_ventilator(None)
-        children_vent = []
-    else:
-        if ventilator_data is not None:
-            ventilator_frequency = variables.get_ventilator_freq()
-            children_vent = utils.add_ventilator_graphs(np.array(ventilator_data), ventilator_frequency)
-            hidden = False
-        else:
-            children_vent = []
-
-    return children_vent, hidden, filename
-
-
-@app.callback(
-    Output({"type": "dynamic-updater", "index": MATCH}, "updateData"),
-    Input({"type": "dynamic-graph", "index": MATCH}, "relayoutData"),
-    State({"type": "dynamic-graph", "index": MATCH}, "id"),
-    prevent_initial_call=True,
-)
-def update_figure(relayoutdata: dict, graph_id_dict: dict):
-    return utils.get_dict(graph_id_dict, relayoutdata)
 
 
 @app.callback(
@@ -127,6 +61,8 @@ def toggle_modal(n1, n2, n3, is_open, selected_file):
     prevent_initial_call=True)
 def get_parent_directory_emg(path_input, stored_cwd, n_clicks, currentdir, path_btn):
     triggered_id = callback_context.triggered_id
+    path = None
+
     if triggered_id == STORED_CWD:
         path = stored_cwd
     elif triggered_id == PATH_BTN:
@@ -153,10 +89,10 @@ def list_cwd_files(cwd):
             is_dir = Path(full_path).is_dir()
             link = html.A([
                 html.Span(
-                file, id={'type': LISTED_FILES, 'index': i},
-                title=full_path,
-                style={'fontWeight': 'bold'} if is_dir else {}
-            )], href='#')
+                    file, id={'type': LISTED_FILES, 'index': i},
+                    title=full_path,
+                    style={'fontWeight': 'bold'} if is_dir else {}
+                )], href='#')
             prepend = '' if not is_dir else '📂'
             cwd_files.append(prepend)
             cwd_files.append(link)
