+++ definitions.py
@@ -44,11 +44,15 @@ default_envelope_value = EnvelopeMethod.RMS
 CONFIRM_CENTERED = 'confirm-centered'
 CWD = 'cwd'
 CWD_FILES = 'cwd-files'
-MODAL_CENTERED = 'modal-centered'
-FILE_PATH_INPUT = 'file-path-input'
+EMG_FREQUENCY_DIV = 'emg-frequency-div'
 EMG_OPEN_CENTERED = 'open-centered-emg'
-VENT_OPEN_CENTERED = 'open-centered-vent'
+EMG_SAMPLING_FREQUENCY = 'emg-sample-freq'
+FILE_PATH_INPUT = 'file-path-input'
+LISTED_FILES = 'listed_file'
+MODAL_CENTERED = 'modal-centered'
+PARENT_DIR = 'parent-dir'
 PATH_BTN = 'path-button'
 STORED_CWD = 'stored-cwd'
-PARENT_DIR = 'parent-dir'
-LISTED_FILES = 'listed_file'
+VENT_FREQUENCY_DIV = 'ventilator-frequency-div'
+VENT_OPEN_CENTERED = 'open-centered-vent'
+VENT_SAMPLING_FREQUENCY = 'ventilator-sample-freq'
